requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
bidict==0.23.1
billiard==4.2.1
blinker==1.9.0
Brotli==1.1.0
celery==5.5.3
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
ConfigArgParse==1.7.1
cryptography==45.0.4
defusedxml==0.7.1
dj-database-url==3.0.1
dj-rest-auth==5.0.2
Django==5.2.3
django-allauth==0.57.0
django-cors-headers==4.7.0
django-debug-toolbar==5.2.0
django-filter==25.1
django-oauth-toolkit==3.0.1
django-redis==6.0.0
django-silk==5.4.0
django-templated-mail==1.1.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
docopt==0.6.2
drf-nested-routers==0.94.2
drf-social-oauth2==3.1.0
exceptiongroup==1.3.0
Flask==3.1.1
flask-cors==6.0.1
Flask-Login==0.6.3
flower==2.0.1
gevent==25.5.1
geventhttpclient==2.3.4
gprof2dot==2025.4.14
greenlet==3.2.3
gunicorn==23.0.0
h11==0.16.0
humanize==4.12.3
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jwcrypto==1.5.6
kombu==5.5.4
locust==2.37.11
locust-cloud==1.24.2
MarkupSafe==3.0.2
model-bakery==1.20.5
msgpack==1.1.1
mysqlclient==2.2.7
oauthlib==3.3.1
packaging==25.0
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
psutil==7.0.0
pycparser==2.22
Pygments==2.19.2
PyJWT==2.9.0
pytest==8.4.1
pytest-django==4.11.1
pytest-mock==3.14.1
pytest-watch==4.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.12.2
python-socketio==5.13.0
python3-openid==3.2.0
pytz==2025.2
pyzmq==27.0.0
razorpay==1.4.2
redis==6.2.0
requests==2.32.4
requests-oauthlib==2.0.0
simple-websocket==1.1.0
six==1.17.0
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
tomli==2.2.1
tornado==6.5.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
watchdog==6.0.0
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.1.3
whitenoise==6.9.0
wsproto==1.2.0
zope.event==5.1
zope.interface==7.2
